Organise specs by policy type
Move specs into namespaces, by policy type
