+++ src/Polly.SharedSpecs/Bulkhead/BulkheadAsyncSpecs.cs
@@ -10,7 +10,7 @@ using Polly.Utilities;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Polly.Specs
+namespace Polly.Specs.Bulkhead
 {
     public class BulkheadAsyncSpecs : BulkheadSpecsHelper
     {
