+++ src/Polly.SharedSpecs/Bulkhead/BulkheadSpecs.cs
@@ -9,7 +9,7 @@ using FluentAssertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Polly.Specs
+namespace Polly.Specs.Bulkhead
 {
     public class BulkheadSpecs : BulkheadSpecsHelper
     {
