+++ src/Polly.SharedSpecs/Bulkhead/BulkheadTResultSpecs.cs
@@ -9,7 +9,7 @@ using FluentAssertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Polly.Specs
+namespace Polly.Specs.Bulkhead
 {
     public class BulkheadTResultSpecs : BulkheadSpecsHelper
     {
