+++ src/Polly.SharedSpecs/CircuitBreaker/CircuitBreakerSpecs.cs
@@ -8,7 +8,7 @@ using Polly.Specs.Helpers;
 using Polly.Utilities;
 using Xunit;
 
-namespace Polly.Specs
+namespace Polly.Specs.CircuitBreaker
 {
     public class CircuitBreakerSpecs : IDisposable
     {
