+++ src/Polly.SharedSpecs/Fallback/FallbackTResultSpecs.cs
@@ -8,7 +8,7 @@ using Xunit;
 
 using Scenario = Polly.Specs.Helpers.PolicyTResultExtensions.ResultAndOrCancellationScenario;
 
-namespace Polly.Specs
+namespace Polly.Specs.Fallback
 {
     public class FallbackTResultSpecs
     {
