+++ src/Polly.SharedSpecs/Retry/RetryForeverAsyncSpecs.cs
@@ -11,7 +11,7 @@ using Xunit;
 
 using Scenario = Polly.Specs.Helpers.PolicyExtensionsAsync.ExceptionAndOrCancellationScenario;
 
-namespace Polly.Specs
+namespace Polly.Specs.Retry
 {
     public class RetryForeverAsyncSpecs
     {
