+++ src/Polly.SharedSpecs/Retry/RetryForeverSpecs.cs
@@ -5,7 +5,7 @@ using FluentAssertions;
 using Polly.Specs.Helpers;
 using Xunit;
 
-namespace Polly.Specs
+namespace Polly.Specs.Retry
 {
     public class RetryForeverSpecs
     {
