+++ src/Polly.SharedSpecs/Retry/RetrySpecs.cs
@@ -6,7 +6,7 @@ using FluentAssertions;
 using Polly.Specs.Helpers;
 using Xunit;
 
-namespace Polly.Specs
+namespace Polly.Specs.Retry
 {
     public class RetrySpecs
     {
