+++ src/Polly.SharedSpecs/Retry/RetryTResultMixedResultExceptionSpecs.cs
@@ -3,7 +3,7 @@ using FluentAssertions;
 using Polly.Specs.Helpers;
 using Xunit;
 
-namespace Polly.Specs
+namespace Polly.Specs.Retry
 {
     public class RetryTResultMixedResultExceptionSpecs
     {
