+++ src/Polly.SharedSpecs/Retry/RetryTResultSpecsAsync.cs
@@ -11,7 +11,7 @@ using Xunit;
 
 using Scenario = Polly.Specs.Helpers.PolicyTResultExtensionsAsync.ResultAndOrCancellationScenario;
 
-namespace Polly.Specs
+namespace Polly.Specs.Retry
 {
     public class RetryTResultSpecsAsync
     {
