+++ src/Polly.SharedSpecs/Retry/WaitAndRetryForeverSpecs.cs
@@ -6,7 +6,7 @@ using Polly.Specs.Helpers;
 using Polly.Utilities;
 using Xunit;
 
-namespace Polly.Specs
+namespace Polly.Specs.Retry
 {
     public class WaitAndRetryForeverSpecs : IDisposable
     {
