+++ src/Polly.SharedSpecs/Timeout/TimeoutSpecs.cs
@@ -7,7 +7,7 @@ using Polly.Timeout;
 using Polly.Utilities;
 using Xunit;
 
-namespace Polly.Specs
+namespace Polly.Specs.Timeout
 {
     public class TimeoutSpecs
     {
